# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "wecom KB API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/auth/login": {
      "get": {
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "returnTo",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/auth/local": {
      "post": {
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 3
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "subject": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string",
                            "enum": [
                              "entra",
                              "paloalto",
                              "local"
                            ]
                          },
                          "email": {
                            "type": "string",
                            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
                            "nullable": true
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "initials": {
                            "type": "string",
                            "maxLength": 2
                          },
                          "active": {
                            "type": "boolean"
                          },
                          "lastLoginAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "roles": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "userId": {
                                  "type": "string",
                                  "format": "uuid"
                                },
                                "roleId": {
                                  "type": "string",
                                  "format": "uuid"
                                },
                                "roleName": {
                                  "type": "string"
                                },
                                "categoryScope": {
                                  "type": "array",
                                  "items": {
                                    "type": "string",
                                    "enum": [
                                      "sim",
                                      "tech",
                                      "billing",
                                      "plans",
                                      "intl",
                                      "ops"
                                    ]
                                  },
                                  "nullable": true
                                },
                                "grantedBy": {
                                  "type": "string",
                                  "format": "uuid",
                                  "nullable": true
                                },
                                "grantedAt": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              },
                              "required": [
                                "userId",
                                "roleId",
                                "roleName",
                                "categoryScope",
                                "grantedBy",
                                "grantedAt"
                              ],
                              "additionalProperties": false
                            }
                          }
                        },
                        "required": [
                          "id",
                          "subject",
                          "source",
                          "email",
                          "displayName",
                          "initials",
                          "active",
                          "lastLoginAt",
                          "roles"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "total": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "items",
                    "total",
                    "page",
                    "pageSize"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "active": {
                    "type": "boolean"
                  },
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "roleId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "categoryScope": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "sim",
                              "tech",
                              "billing",
                              "plans",
                              "intl",
                              "ops"
                            ]
                          },
                          "nullable": true
                        }
                      },
                      "required": [
                        "roleId",
                        "categoryScope"
                      ],
                      "additionalProperties": false
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "auditId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "auditId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}